+++ homeassistant/components/vesync/__init__.py
@@ -5,18 +5,18 @@ import logging
 from pyvesync import VeSync
 from pyvesync.utils.errors import VeSyncLoginError
 
-from homeassistant.config_entries import ConfigEntry, ConfigEntryState
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME, Platform
-from homeassistant.core import HomeAssistant, ServiceCall
-from homeassistant.exceptions import ConfigEntryAuthFailed, ServiceValidationError
+from homeassistant.core import HomeAssistant
+from homeassistant.exceptions import ConfigEntryAuthFailed
 from homeassistant.helpers import config_validation as cv, entity_registry as er
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.device_registry import DeviceEntry
-from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.helpers.typing import ConfigType
 
-from .const import DOMAIN, SERVICE_UPDATE_DEVS, VS_COORDINATOR, VS_MANAGER
+from .const import DOMAIN, VS_COORDINATOR, VS_MANAGER
 from .coordinator import VeSyncDataCoordinator
+from .services import async_setup_services
 
 CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)
 
@@ -33,6 +33,14 @@ PLATFORMS = [
 ]
 
 _LOGGER = logging.getLogger(__name__)
+
+
+async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
+    """Set up my integration."""
+
+    async_setup_services(hass)
+
+    return True
 
 
 async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
@@ -75,35 +83,6 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data.pop(DOMAIN)
 
     return unload_ok
-
-
-async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
-    """Set up my integration."""
-
-    async def async_new_device_discovery(service: ServiceCall) -> None:
-        """Discover and add new devices."""
-
-        entries = hass.config_entries.async_entries(DOMAIN)
-        entry = entries[0] if entries else None
-
-        if not entry:
-            raise ServiceValidationError("Entry not found")
-        if entry.state is not ConfigEntryState.LOADED:
-            raise ServiceValidationError("Entry not loaded")
-        manager = hass.data[DOMAIN][VS_MANAGER]
-        known_devices = list(manager.devices)
-        await manager.get_devices()
-        new_devices = [
-            device for device in manager.devices if device not in known_devices
-        ]
-
-        if new_devices:
-            async_dispatcher_send(hass, "vesync_new_devices", new_devices)
-
-    hass.services.async_register(
-        DOMAIN, SERVICE_UPDATE_DEVS, async_new_device_discovery
-    )
-    return True
 
 
 async def async_migrate_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
